psxtests, psxtmtests: Address -Wsign-compare warnings
Most of these warnings were between int or ssize_t and size_t.
In particular, various POSIX system calls like read() and write()
return ssize_t and comparing that value with the sizeof a buffer
is a common source of these warnings. Another common source is
using an int as the iterator in a for loop with the limit being
a size_t.

With the type change, some printf() specifiers needed to change also.

+++ testsuites/psxtests/psxhdrs/time/clock.c
@@ -45,5 +45,5 @@ int test( void )
 
   clock_ticks = clock();
 
-  return (clock_ticks != -1);
+  return ((int) clock_ticks != -1);
 }
